1.1.0-1.17.1:     - Fixed creative items having to large default trace_width     - Battery level and trace_width now are stored as int and so can reach 2147483647 (2^31-1)     - Changed license field within the mods.toml and added a tiny description

+++ src/main/java/de/canitzp/justabattery/BatteryCombiningRecipe.java
@@ -57,12 +57,12 @@ public class BatteryCombiningRecipe extends ShapelessRecipe {
         return combinedLevel;
     }
     
-    private byte getMaxTraceWidth(CraftingContainer inv){
-        byte combinedTraceWidth = 0;
+    private int getMaxTraceWidth(CraftingContainer inv){
+        int combinedTraceWidth = 0;
         for(int i = 0; i < inv.getContainerSize(); i++){
             ItemStack stack = inv.getItem(i);
             if(!stack.isEmpty() && stack.getItem() instanceof BatteryItem){
-                byte traceWidth = BatteryItem.getTraceWidth(stack);
+                int traceWidth = BatteryItem.getTraceWidth(stack);
                 if(traceWidth > combinedTraceWidth){
                     combinedTraceWidth = traceWidth;
                 }
